+++ tests/subsys/fs/fs_api/src/test_fs.c
@@ -11,6 +11,7 @@
 #include <errno.h>
 #include <init.h>
 #include <fs/fs.h>
+#include <fs/fs_sys.h>
 #include <sys/__assert.h>
 #include "test_fs.h"
